Skipping level 2
Cloned the entire level 1 and sat up the exact same settings as the
original and moved it to meet with level 3.
Deleted the redundant loadlevel3 and levelloader from the second level 1
though.
Checking the skip will make the second level 1 active and leaving it
blank will  have the original level 1 active. The other one will be
deactivated immediadly upon playing the level.